+++ foxes/input/states/single.py
@@ -29,15 +29,15 @@ class SingleStateStates(States):
 
     """
 
-    def __init__(self, ws, wd, ti=None, rho=None, profiles={}, **profdata):
+    def __init__(self, ws=None, wd=None, ti=None, rho=None, profiles={}, **profdata):
         """
         Constructor.
 
         Parameters
         ----------
-        ws: float
+        ws: float, optional
             The wind speed
-        wd: float
+        wd: float, optional
             The wind direction
         ti: float, optional
             The TI value
@@ -57,6 +57,9 @@ class SingleStateStates(States):
         self.rho = rho
         self.profdicts = profiles
         self.profdata = profdata
+
+        if ws is None and wd is None and ti is None and rho is None and not len(profiles):
+            raise KeyError(f"Expecting at least one parameter: ws, wd, ti, rho, profiles")
 
     def initialize(self, algo, verbosity=0):
         """
@@ -129,16 +132,18 @@ class SingleStateStates(States):
             The output variable names
 
         """
-        out = []
+        out = set()
         if self.ws is not None:
-            out.append(FV.WS)
+            out.add(FV.WS)
         if self.wd is not None:
-            out.append(FV.WD)
+            out.add(FV.WD)
         if self.ti is not None:
-            out.append(FV.TI)
+            out.add(FV.TI)
         if self.rho is not None:
-            out.append(FV.RHO)
-        return out
+            out.add(FV.RHO)
+        out.update(list(self._profiles.keys()))
+
+        return list(out)
 
     def weights(self, algo):
         """
@@ -200,9 +205,8 @@ class SingleStateStates(States):
                 (pdata.n_states, pdata.n_points), self.rho, dtype=FC.DTYPE
             )
 
-        z = pdata[FC.POINTS][:, :, 2]
         if len(self._profiles):
-            z = pdata[FC.POINTS][:, :, 2]
+            z = pdata[FC.POINTS][:, :, 2]          
             for k, v in self.profdata.items():
                 pdata[k] = v
             for v, p in self._profiles.items():